+++ src/main/scala/de/uni_luebeck/isp/tessla/tessladoc/Main.scala
@@ -11,18 +11,20 @@ object Main extends SexyOpt {
   override val version = Some(BuildInfo.version)
   override val programDescription = "Generate documentation for TeSSLa code"
 
-  val files = restArgs("file", "The TeSSLa files for which to generate documentation", atLeastOne = false)
-  //val stdLib = flag("stdlib", 's', "Include documentation for definitions from the standard library")
+  val files = restArgs("files", "The TeSSLa files for which to generate documentation", atLeastOne = false)
+  val stdLib = flag("stdlib", 's', "Include documentation for definitions from the standard library")
   val includes = flag("includes", 'i', "Include documentation from included files")
+  val globalsOnly = flag("globals-only", 'g', "Do not show information for local definitions")
 
   def main(args: Array[String]) = {
     parse(args)
     val streams = files.map(CharStreams.fromFileName)
     val includeResolver = optionIf(includes)(IncludeResolvers.fromFile _)
-    TesslaDoc.extract(streams, includeResolver) match {
+    TesslaDoc.extract(streams, includeResolver, includeStdlib = stdLib) match {
       case Success(tesslaDocs, warnings) =>
         warnings.foreach(w => System.err.println(s"Warning: $w"))
-        println(tesslaDocs)
+        val relevantDocs = if(globalsOnly) tesslaDocs.globalsOnly else tesslaDocs
+        println(relevantDocs)
       case Failure(errors, warnings) =>
         warnings.foreach(w => System.err.println(s"Warning: $w"))
         errors.foreach(e => System.err.println(s"Error: $e"))