+++ src/main/scala/de/uni_luebeck/isp/tessla/tessladoc/TesslaDoc.scala
@@ -1,12 +1,15 @@
 package de.uni_luebeck.isp.tessla.tessladoc
 
-import de.uni_luebeck.isp.tessla.TranslationPhase.Result
+import de.uni_luebeck.isp.tessla.Errors.InternalError
+import de.uni_luebeck.isp.tessla.TranslationPhase.{Failure, Result, Success}
 import de.uni_luebeck.isp.tessla._
 import org.antlr.v4.runtime.{CharStream, ParserRuleContext, Token}
 
 import scala.collection.JavaConverters._
 
-sealed abstract class TesslaDoc extends TesslaDoc.DocElement
+sealed abstract class TesslaDoc extends TesslaDoc.DocElement {
+  def isGlobal: Boolean
+}
 
 object TesslaDoc {
   trait DocElement {
@@ -18,6 +21,10 @@ object TesslaDoc {
 
   case class ModuleDoc(items: Seq[TesslaDoc]) extends TesslaDoc {
     override def toJSON = items.mkString("[\n", ",\n", "\n]")
+
+    override def isGlobal = true
+
+    def globalsOnly = ModuleDoc(items.filter(_.isGlobal))
   }
 
   case class TypeDoc(name: String, typeParameters: Seq[String], doc: String, loc: Location) extends TesslaDoc {
@@ -31,6 +38,8 @@ object TesslaDoc {
          |  "doc": "$doc"
          |}""".stripMargin
     }
+
+    override def isGlobal = true
   }
 
   case class Param(name: String, typ: String) extends DocElement {
@@ -68,6 +77,8 @@ object TesslaDoc {
       returnType.foreach(typ => str += s"""  "returnType": "$typ",\n""")
       str + s"""  "doc": "$doc"\n}"""
     }
+
+    override def isGlobal = scope == Global
   }
 
   class Extractor(spec: Seq[TesslaParser.ParseResult]) extends TranslationPhase.Translator[ModuleDoc] {
@@ -131,13 +142,29 @@ object TesslaDoc {
     }
   }
 
-  def extract(srcs: Seq[CharStream], includeResolver: Option[String => Option[CharStream]]) = {
+  def extract(srcs: Seq[CharStream], includeResolver: Option[String => Option[CharStream]], includeStdlib: Boolean): Result[ModuleDoc] = {
     Result.runSequentially(srcs) { src =>
       val results =
         includeResolver.map(new TesslaParser.WithIncludes(_).translate(src)).getOrElse {
           TesslaParser.SingleFile.translate(src).map(Seq(_))
         }
       results.andThen(new Extractor(_).translate())
+    }.andThen { docsForFiles =>
+      if(includeStdlib) {
+        forStdlib.map { docsForStdlib =>
+          docsForStdlib +: docsForFiles
+        }
+      } else {
+        Success(docsForFiles, Seq())
+      }
     }.map(modules => ModuleDoc(modules.flatMap(_.items)))
+  }
+
+  def forStdlib: Result[ModuleDoc] = {
+    IncludeResolvers.fromStdlibResource("Predef.tessla").map { predef =>
+      extract(Seq(predef), Some(IncludeResolvers.fromStdlibResource), includeStdlib = false)
+    }.getOrElse {
+      Failure(Seq(InternalError("Could not find standard library")), Seq())
+    }
   }
 }